+++ .gitignore
@@ -7,3 +7,4 @@ test_model/
 Images/
 scratch/
 hypers/
+atari_prediction/policies/
