+++ age_estimation/main/HL_tuner.py
@@ -1,14 +1,12 @@
-import tensorflow as tf
 from tensorflow import keras
-from tensorflow.keras import layers
 import keras_tuner as kt
 from experiment.hypermodels import HyperRegression, HyperHLGaussian, HyperHLOneBin
 import os
 import sys
 import json
-from experiment.datasets import MegaAgeDataset, UTKFaceDataset
+from age_estimation.datasets import UTKFaceDataset
 from experiment.logging import LogGridSearch
-from experiment.get_model import get_model
+from age_estimation.base_models import get_model
 
 
 def main(base_dir):
