+++ age_estimation/main/fgnet.py
@@ -1,14 +1,12 @@
-import tensorflow as tf
 from tensorflow import keras
-from tensorflow.keras import layers
 import keras_tuner as kt
 from experiment.hypermodels import HyperRegression, HyperHLGaussian, HyperHLOneBin
 import os
 import sys
 import json
-from experiment.datasets import FGNetDataset
+from age_estimation.datasets import FGNetDataset
 from experiment.logging import LogGridSearch
-from experiment.get_model import get_model
+from age_estimation.base_models import get_model
 from experiment.preprocessing import Scaler
