+++ age_estimation/main/main.py
@@ -1,8 +1,8 @@
 import tensorflow as tf
 from tensorflow import keras
-from tensorflow.keras import layers
+from keras import layers
 from experiment.models import HLGaussian, HLOneBin, Regression
-from experiment.datasets import MegaAgeDataset
+from age_estimation.datasets import MegaAgeDataset
 import os
 import sys
 import json
