+++ age_estimation/scripts/test_run.sh
