+++ age_estimation/scripts/worker.sh
